fix: wire trip-planning Redis lifecycle
