+++ services/trip-planning/src/trip_planning/adapters/messaging/__init__.py
@@ -1,7 +1,9 @@
 from .publisher import RedisEventPublisher
-from .subscriber import RedisEventSubscriber
+from .redis_streams import RedisEventSubscriber
+from .subscriber import start_trip_planning_subscription
 
 __all__ = [
     "RedisEventPublisher",
     "RedisEventSubscriber",
+    "start_trip_planning_subscription",
 ]
